+++ package.json
@@ -18,7 +18,9 @@
     "streams:write": "node src/streams/write.js",
     "streams:transform": "node src/streams/transform.js",
     "zip:compress": "node src/zip/compress.js",
-    "zip:decompress": "node src/zip/decompress.js"
+    "zip:decompress": "node src/zip/decompress.js",
+    "modules": "node src/modules/cjsToEsm.mjs"
+
   },
   "repository": {
     "type": "git",
